% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{analyticGaussianMechanism}
\alias{analyticGaussianMechanism}
\title{Calculated sigma for analytic Gaussian mechanism}
\usage{
analyticGaussianMechanism(epsilon, delta, sens, tol = 1e-12)
}
\arguments{
\item{epsilon}{(`numeric()`) Epsilon > 0}

\item{delta}{(`numeric()`) Delta that is between (0,1)}

\item{sens}{(`numeric()`) Sensitivity of the algorithm}

\item{tol}{(`numeric()`) Tolerance for binary search}
}
\value{
(`numeric()`) Sigma that can be used to generate calibrated Gaussian noise
}
\description{
Caluclates the minimal valid sigma for given privacy parameters for Gaussian mechanism
  (see https://arxiv.org/abs/1805.06530).
}
\author{
Raphael Rehms
}
