% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictModel.R
\name{assignPredictModel}
\alias{assignPredictModel}
\title{Wrapper to predict model on server site}
\usage{
assignPredictModel(bin_call)
}
\arguments{
\item{bin_call}{(`character(1L)`) Binary predict call (encodeObject applied on call).}
}
\value{
Vector of predictions
}
\description{
This function enables to make predictions on the server for any
  model.
}
\author{
Daniel S.
}
