% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm_ci.R
\name{aucCI}
\alias{aucCI}
\title{Calculate CI for AUC}
\usage{
aucCI(connections, roc_glm, alpha = 0.05)
}
\arguments{
\item{connections}{(`DSI::connection`) Connection to an OPAL server.}

\item{roc_glm}{(`list()`) List containing the ROC-GLM parameter returned from `dsROCGLM`.}

\item{alpha}{(`numeric(1L)`) Significance level alpha (default is `0.05`).}
}
\value{
Numeric vector with two values containing the boundaries of the confidence interval.
}
\description{
This function calculates a CI for the AUC based on the approach proposed by DeLong.
}
\author{
Daniel S.
}
