% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier_score.R
\name{brierScore}
\alias{brierScore}
\title{Calculate Brier score}
\usage{
brierScore(truth_name, prob_name)
}
\arguments{
\item{truth_name}{(`character(1L)`) Character containing the name of the vector of 0-1-values
encoded as integer or numeric.}

\item{prob_name}{(`character(1L)`) Character containing the name of the vector of probabilities.}
}
\value{
Brier score and number of values used as weights for later aggregation.
}
\description{
This function calculates the Brier score for a given 0-1-vector and
  vector of probabilities.
}
\author{
Daniel S.
}
