% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm_parts.R
\name{calculateLambda}
\alias{calculateLambda}
\title{Transform Response for Probit Fisher-Scoring}
\usage{
calculateLambda(y, X, beta)
}
\arguments{
\item{y}{Original 0-1-response}

\item{X}{Design matrix}

\item{beta}{Estimated parameter vector}
}
\value{
Numeric vector of new response
}
\description{
This function transform the original 0-1-response y  to a new reponse
  "lambda" that is used as respone for the fisher scoring.
}
\author{
Stefan B., Daniel S.
}
