% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve.R
\name{calibrationCurve}
\alias{calibrationCurve}
\title{Calculate data for calibration curve}
\usage{
calibrationCurve(
  truth_name,
  prob_name,
  nbins = 10L,
  remove_critical_bins = TRUE
)
}
\arguments{
\item{truth_name}{(`character(1L)`) Character containing the name of the vector of
0-1-values encoded as integer or numeric.}

\item{prob_name}{(`character(1L)`) Character containing the name of the vector of probabilities.}

\item{nbins}{(`integer(1L)`) Number of bins between 0 and 1 (default is 10).}

\item{remove_critical_bins}{(`logical(1L)`) Just remove critical bins and return just bins with `length(bin) > n_critical`.}
}
\value{
Brier score and number of values used as weights for later aggregation.
}
\description{
This function calculates a vector of values for the calibration curve.
}
\author{
Daniel S.
}
