% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm_parts.R
\name{checkTruthProb}
\alias{checkTruthProb}
\title{Truth and Prediction Checker}
\usage{
checkTruthProb(truth_name, prob_name, pos = NULL)
}
\arguments{
\item{truth_name}{[character(1L)] Character containing the name of the vector of 0-1-values encoded as integer or numeric.}

\item{prob_name}{[character(1L)] Character containing the name of the vector of probabilities.}

\item{pos}{[character(1L)] Character containing the name of the positive class.}
}
\value{
Data frame containing the truth and prediction vector.
}
\description{
This function checks if the vector of true values and predictions
  has the correct format to be used for the ROC-GLM. If something does not suit,
  then the function tries to convert it into the correct format.
}
\author{
Daniel S.
}
