% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm_parts.R
\name{computePlacementValues}
\alias{computePlacementValues}
\title{Compute Placement Values on Server}
\usage{
computePlacementValues(pooled_scores_name, truth_name, prob_name, tset = NULL)
}
\arguments{
\item{pooled_scores_name}{[character(1L)] Name of the object holding the pooled negative scores.}

\item{truth_name}{[character(1L)] Character containing the name of the vector of 0-1-values encoded as integer or numeric.}

\item{prob_name}{[character(1L)] Character containing the name of the vector of probabilities.}

\item{tset}{Set of thresholds}
}
\value{
Placement values
}
\description{
This function calculates the placement values required to calculate the ROC-GLM.
}
\author{
Stefan B., Daniel S.
}
