% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2sens.R
\name{dsL2Sens}
\alias{dsL2Sens}
\title{Calculate the l2 sensitivity on DataSHIELD servers}
\usage{
dsL2Sens(
  connections,
  dat_name,
  pred_name,
  nbreaks = NULL,
  cols = NULL,
  drop_on_error = TRUE
)
}
\arguments{
\item{connections}{(`DSI::connection`) Connection to an OPAL server.}

\item{dat_name}{(`character(1)`) The name of the data at the DataSHIELD servers.}

\item{pred_name}{(`character(1)`) Name of the prediction object at the DataSHIELD server.}

\item{nbreaks}{(`integer(1L)`) Number of breaks used for the histogram
(default = nrow(dat) / 3).}

\item{cols}{(`character()`) Subset of columns used to find adjacent inputs.}

\item{drop_on_error}{(`logical(1L)`) NA is returned if an error occurs on a server.}
}
\value{
List with maximal l2 sensitivity, indices of inputs that are
  used to calculate the maximal l2 sensitivity, and the number of adjacent inputs.
}
\description{
Calculation of the l2 sensitivity using a histogram representetation.
  Source: https://www.cis.upenn.edu/~aaroth/Papers/privacybook.pdf
}
\author{
Daniel S.
}
