% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{erf}
\alias{erf}
\title{Calculates the error function used for the analytic Gaussian mechanism}
\usage{
erf(x)
}
\arguments{
\item{x}{(`numeric()`) A (vector of) real number(s)}
}
\value{
(`numeric()`) The evaluated error function
}
\description{
Calculates the error function used for the analytic Gaussian mechanism
}
\author{
Raphael Rehms
}
