% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{generateParameterTableDP}
\alias{generateParameterTableDP}
\title{Generate all combinations of DP parameters fromt he paper}
\usage{
generateParameterTableDP()
}
\value{
data.frame with combinations and info whether parameters are valid accorsing to the paper.
}
\description{
This function creates a data frame with the cosidered values from the paper. One may want to choose or set a better grid or method in the future.
}
\author{
Raphael R.
}
