% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm_parts.R
\name{getPositiveScores}
\alias{getPositiveScores}
\title{Return positive scores}
\usage{
getPositiveScores(truth_name, prob_name)
}
\arguments{
\item{truth_name}{(`character(1L)`) Character containing the name of the vector of 0-1-values
encoded as integer or numeric.}

\item{prob_name}{(`character(1L)`) Character containing the name of the vector of probabilities.}
}
\value{
Positive scores
}
\description{
This function just returns positive scores and is used
  as aggregator to send these positive scores.
}
\author{
Daniel S.
}
