% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_curve.R
\name{plot.calibration.curve}
\alias{plot.calibration.curve}
\title{Plot calibration curve}
\usage{
\method{plot}{calibration.curve}(x, ...)
}
\arguments{
\item{x}{(`list()`) Object returned by `dsCalibrationCurve()`}

\item{...}{Additional arguments passed to `plotCalibrationCurve()`.}
}
\value{
ggplot of calibration curve(s)
}
\description{
This function plots the calibration curve returned by `dsCalibrationCurve()`.
}
\author{
Daniel S.
}
