% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrationCurve.R
\name{plotCalibrationCurve}
\alias{plotCalibrationCurve}
\title{Plot calibration curve}
\usage{
plotCalibrationCurve(cc, individuals = TRUE, ...)
}
\arguments{
\item{cc}{(`list()`) Object returned by `dsCalibrationCurve()`}

\item{individuals}{(`logical(1L)`) Logical value indicating whether the individual calibration
curves should be plotted or not (default is `TRUE`).}

\item{...}{Additional arguments passed to `geom_point()` and `geom_line()` for the calibration line and points.}
}
\value{
ggplot of calibration curve(s)
}
\description{
This function plots the calibration curve returned by `dsCalibrationCurve()`.
}
\author{
Daniel S.
}
