% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm_parts.R
\name{probitLikelihood}
\alias{probitLikelihood}
\title{Calculate Likelihood of Probit Model}
\usage{
probitLikelihood(y, X, beta, w = NULL)
}
\arguments{
\item{y}{Original 0-1-response}

\item{X}{Design matrix}

\item{beta}{Estimated parameter vector}

\item{x}{Weight vector}
}
\value{
Numeric value containing the likelihood
}
\description{
This function calculates the likelihood used in a probit regression (Bernoulli distribution + probit link).
}
\author{
Stefan B., Daniel S.
}
