% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_object.R
\name{pushObject}
\alias{pushObject}
\title{Push object to DataSHIELD server}
\usage{
pushObject(
  connections,
  obj,
  check_serialization = TRUE,
  package = NULL,
  just_return_call = FALSE
)
}
\arguments{
\item{connections}{(`DSI::connection`) Connection to an OPAL server.}

\item{obj}{(arbitrary R object) Object containing which should be pushed to the server.}

\item{check_serialization}{(`logical(1L)`) Check if the serialized model can be deserialized
locally (default is `TRUE`).}

\item{package}{(`character(1L)`) Package required for model predictions (default is `NULL`).}

\item{just_return_call}{(`logical(1L)`) Just return the call and not execute on server (mainly for testing purposes,
default is `FALSE`).}
}
\description{
This function pushes a serialized object and assigns a server
  variable to it.
}
\examples{
# Just for local testing:
cl = pushObject("Dummy", iris, just_return_call = TRUE)
all.equal(iris, eval(parse(text = cl)))
}
\author{
Daniel S.
}
