% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_missings.R
\name{removeMissings}
\alias{removeMissings}
\title{Remove missing values from data}
\usage{
removeMissings(symbol)
}
\arguments{
\item{symbol}{(`character(1L)`) R object from which missings should be removed.}
}
\value{
Object defined in symbol wrapped by `na.omit`.
}
\description{
Remove missing values from data and assign it to new data or old data.
}
\author{
Daniel S.
}
