% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm_parts.R
\name{rocGLMData}
\alias{rocGLMData}
\title{Get Data for ROC-GLM}
\usage{
rocGLMData(U, tset)
}
\arguments{
\item{U}{(`matrix()`) Response matrix for ROC-GLM}

\item{tset}{(`numeric()`) Set of thresholds}
}
\value{
Data used for the ROC-GLM
}
\description{
This function calculates the data used for the ROC-GLM
}
\author{
Stefan B., Daniel S.
}
