% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm_parts.R
\name{rocGLMFrame}
\alias{rocGLMFrame}
\title{Calculate data for ROC-GLM}
\usage{
rocGLMFrame(truth_name, prob_name, pooled_scores_name)
}
\arguments{
\item{truth_name}{(`character(1L)`) Character containing the name of the vector of 0-1-values
encoded as integer or numeric.}

\item{prob_name}{(`character(1L)`) Character containing the name of the vector of probabilities.}

\item{pooled_scores_name}{(`character(1L)`) Character containing the name of the object holding
the pooled negative scores.}
}
\value{
Data.frame used to calculate ROC-GLM
}
\description{
This function stores the data on the server used for the ROC-GLM
}
\author{
Stefan B., Daniel S.
}
