% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm_ci.R
\name{aucCI}
\alias{aucCI}
\title{Calculate CI for AUC}
\usage{
aucCI(
  connections,
  truth_name,
  pred_name,
  roc_glm,
  alpha = 0.05,
  epsilon = 0.2,
  delta = 0.2
)
}
\arguments{
\item{connections}{(`DSI::connection`) Connection to an OPAL server.}

\item{truth_name}{(`character(1L)`) Character containing the name of the vector of 0-1-values
encoded as integer or numeric.}

\item{pred_name}{(`character(1L)`) Character containing the name of the vector of probabilities.}

\item{roc_glm}{(`list()`) List containing the ROC-GLM parameter returned from `dsROCGLM`.}

\item{alpha}{(`numeric(1L)`) Significance level alpha (default is `0.05`).}

\item{epsilon}{(`numeric(1L)`) Privacy parameter for differential privacy (DP).}

\item{delta}{(`numeric(1L)`) Probability of violating epsilon DP.}
}
\value{
Numeric vector with two values containing the boundaries of the confidence interval.
}
\description{
This function calculates a CI for the AUC based on the approach proposed by DeLong.
}
\author{
Daniel S.
}
