% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm.R
\name{dsROCGLM}
\alias{dsROCGLM}
\title{Calculate Probit Regression on Server}
\usage{
dsROCGLM(
  connections,
  truth_name,
  pred_name,
  trace = TRUE,
  clean_server = TRUE,
  alpha = 0.05
)
}
\arguments{
\item{connections}{[DSI::connection] Connection to an OPAL server.}

\item{truth_name}{[character(1L)] Character containing the name of the vector of 0-1-values encoded as integer or numeric.}

\item{trace}{[logical(1L)] If `TRUE` (default), information about the progress is shown.}

\item{clean_server}{[logical(1L)] Set to `TRUE` (default) if all temprary data stored on the server should
be removed when the fitting is finished.}

\item{alpha}{[numeric(1L)] Significance level alpha for confidence interval (default is `0.05`).}

\item{prob_name}{[character(1L)] Character containing the name of the vector of probabilities.}
}
\value{
List with estimated parameter, number of iterations, and the deviance when the algorithm is stopped.
}
\description{
ROC-GLM function for DataSHIELD. The function prepares the data on the server and fits the model.
  The model returned by the function contains ROC-GLM as parametrized model.
}
\author{
Daniel S.
}
