% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{getNegativeScoresVar}
\alias{getNegativeScoresVar}
\title{Return variance of negative scores}
\usage{
getNegativeScoresVar(truth_name, prob_name, m = NULL, return_sum = FALSE)
}
\arguments{
\item{truth_name}{(`character(1L)`) Character containing the name of the vector of 0-1-values
encoded as integer or numeric.}

\item{prob_name}{(`character(1L)`) Character containing the name of the vector of probabilities.}

\item{m}{(`numeric(1L)`) Sample mean used for variance calculation. If `NULL` (default), the
sample mean of the negative scores is used.}

\item{return_sum}{(`logical(1L)`) Logical value indicating whether the function should
just return the sum of negative scores.}
}
\value{
Variance of differences of Negative scores
}
\description{
This function just returns the variance of negative scores.
}
\author{
Daniel S.
}
