% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{getPositiveScores}
\alias{getPositiveScores}
\title{Return positive scores}
\usage{
getPositiveScores(truth_name, prob_name, sd_noise, ntimes = 2L)
}
\arguments{
\item{truth_name}{(`character(1L)`) Character containing the name of the vector of 0-1-values
encoded as integer or numeric.}

\item{prob_name}{(`character(1L)`) Character containing the name of the vector of probabilities.}

\item{sd_noise}{(`numerical(1L)`) Standard deviation used to add noise to the return value.}

\item{ntimes}{(`integer(1L)`) Times the standard deviation used for simulating noise added to the data.}
}
\value{
Positive scores
}
\description{
This function just returns positive scores and is used
  as aggregator to send these positive scores.
}
\author{
Daniel S.
}
