% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{getPositiveScores}
\alias{getPositiveScores}
\title{Return positive scores}
\usage{
getPositiveScores(
  truth_name,
  prob_name,
  epsilon = 0.2,
  delta = 0.2,
  seed_object = NULL,
  sort = FALSE
)
}
\arguments{
\item{truth_name}{(`character(1L)`) Character containing the name of the vector of 0-1-values
encoded as integer or numeric.}

\item{prob_name}{(`character(1L)`) Character containing the name of the vector of probabilities.}

\item{epsilon}{(`numeric(1L)`) Privacy parameter for differential privacy (DP).}

\item{delta}{(`numeric(1L)`) Probability of violating epsilon DP.}

\item{seed_object}{(`character(1L)`) Name of an object which is used
to add a seed based on an object.}
}
\value{
Positive scores
}
\description{
This function just returns positive scores and is used
  as aggregator to send these positive scores.
}
\author{
Daniel S.
}
