% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{getPositiveScoresVar}
\alias{getPositiveScoresVar}
\title{Return variance of score differneces}
\usage{
getPositiveScoresVar(truth_name, prob_name)
}
\arguments{
\item{truth_name}{(`character(1L)`) Character containing the name of the vector of 0-1-values
encoded as integer or numeric.}

\item{prob_name}{(`character(1L)`) Character containing the name of the vector of probabilities.}
}
\value{
Variance of differences of positive scores
}
\description{
Return variance of score differneces
}
\author{
Daniel S.
}
