% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{GMVar}
\alias{GMVar}
\title{Calculate standard deviation for Gaussian Mechanism}
\usage{
GMVar(l2s, epsilon, delta)
}
\arguments{
\item{l2s}{(`numeric(1L)`) l2-sensitivity.}

\item{epsilon}{(`numeric(1L)`) First privacy parameter for (e,d)-differential privacy.}

\item{delta}{(`numeric(1L)`) second privacy parameter for (e,d)-differential privacy.}
}
\value{
Numerical value for the standard deviation for the normal distribution.
}
\description{
Calculate standard deviation for Gaussian Mechanism
}
\author{
Daniel S.
}
