% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l2sens.R
\name{l2sens}
\alias{l2sens}
\title{Calculate the l2 sensitivity}
\usage{
l2sens(dat, scores, nbreaks = NULL, cols = NULL, norm = diff)
}
\arguments{
\item{dat}{(`data.frame()`) The data used to find adjacent inputs.}

\item{scores}{(`numeric()`) Predicted scores/probabilities.}

\item{nbreaks}{(`integer(1L)`) Number of breaks used for the histogram
(default = nrow(dat) / 3).}

\item{cols}{(`character()`) Subset of columns used to find adjacent inputs.}

\item{norm}{(`function()`) Function to calculate the differnece between the
scores of adjacent inputs (default = Euclidean norm).}
}
\value{
List with maximal l2 sensitivity, indices of inputs that are
  used to calculate the maximal l2 sensitivity, and the number of adjacent inputs.
}
\description{
Calculation of the l2 sensitivity using a histogram representetation.
  Source: https://www.cis.upenn.edu/~aaroth/Papers/privacybook.pdf
}
\author{
Daniel S.
}
