% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_glm_print.R
\name{print.ROC.GLM}
\alias{print.ROC.GLM}
\title{Printer for ROC-GLM}
\usage{
\method{print}{ROC.GLM}(x, ...)
}
\arguments{
\item{x}{(`list()`) List containing the ROC-GLM parameter returned from `dsROCGLM`.}

\item{...}{Additional parameters (basically none, but CRAN forces us to do this in print).}
}
\description{
This function provides a printer for the ROC-GLM object returned from `dsROCGLM`.
}
\author{
Daniel S.
}
