% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{seedBoundedToObject}
\alias{seedBoundedToObject}
\title{Get a seed depending on an object}
\usage{
seedBoundedToObject(object)
}
\arguments{
\item{object}{(`character(1L)`) Character containing the name of the object
to which the seed is bounded.}
}
\value{
Integer containing a seed.
}
\description{
This function creates a seed based on the hash of an object.
}
\author{
Daniel S.
}
